requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2

# Database
psycopg2-binary==2.9.9

# Cache & Broker
redis==5.0.4
django-redis==5.4.0
celery==5.4.0
kombu==5.3.7

# AI
openai==1.30.1
tiktoken==0.7.0

# Rate limiting
django-ratelimit==4.1.0
django-celery-beat==2.6.0

# Monitoring & Logging
django-extensions==3.2.3
flower==2.0.1

# Utils
python-decouple==3.8
Pillow==10.3.0
requests==2.32.2
python-dateutil==2.9.0

# Production
gunicorn==22.0.0
whitenoise==6.7.0

# Testing
pytest-django==4.8.0
pytest-asyncio==0.23.7
factory-boy==3.3.0

dj-database-url==2.1.0